Move DirectoryTraversal into its own module
In preparation for adding functionality for ignoring files/dirs in the
project.
